Update formatting, split tests, some spacing
